Correct count("SKILLSIT") to return a Double rather than an integer
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@23684 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/util/JepCountType.java
@@ -838,7 +838,7 @@ public abstract class JepCountType
 				count++; //For the skill
 				count += sk.getSizeOfListFor(ListKey.SITUATION);
 			}
-			return Integer.valueOf(count);
+			return Double.valueOf(count);
 		}
 		
 	}
